+++ console/settings.py
@@ -2,10 +2,11 @@ import logging
 import os
 
 from sdx.common.logger_config import logger_initial_config
+from structlog import wrap_logger
 
 
 logger_initial_config(service_name='sdx-console')
-logger = logging.getLogger(__name__)
+logger = wrap_logger(logging.getLogger(__name__))
 
 DB_HOST = os.getenv('POSTGRES_HOST', '0.0.0.0')
 DB_PORT = os.getenv('POSTGRES_PORT', '5432')
@@ -29,3 +30,37 @@ HB_INTERVAL = os.getenv("HB_INTERVAL", 30)
 
 SDX_DECRYPT_URL = os.getenv("SDX_DECRYPT_URL", "http://sdx-decrypt:5000/decrypt")
 SDX_STORE_URL = os.getenv("SDX_STORE_URL", "http://sdx-store:5000/")
+
+RABBIT_URL = 'amqp://{user}:{password}@{hostname}:{port}/{vhost}'.format(
+    hostname=os.getenv('RABBITMQ_HOST', 'rabbit'),
+    port=os.getenv('RABBITMQ_PORT', 5672),
+    user=os.getenv('RABBITMQ_DEFAULT_USER', 'rabbit'),
+    password=os.getenv('RABBITMQ_DEFAULT_PASS', 'rabbit'),
+    vhost=os.getenv('RABBITMQ_DEFAULT_VHOST', '%2f')
+)
+RABBIT_URL2 = 'amqp://{user}:{password}@{hostname}:{port}/{vhost}'.format(
+    hostname=os.getenv('RABBITMQ_HOST2', 'rabbit'),
+    port=os.getenv('RABBITMQ_PORT2', 5672),
+    user=os.getenv('RABBITMQ_DEFAULT_USER', 'rabbit'),
+    password=os.getenv('RABBITMQ_DEFAULT_PASS', 'rabbit'),
+    vhost=os.getenv('RABBITMQ_DEFAULT_VHOST', '%2f')
+)
+RABBIT_URLS = [RABBIT_URL, RABBIT_URL2]
+RABBIT_SURVEY_QUEUE = os.getenv('RABBIT_SURVEY_QUEUE', 'survey')
+
+
+def get_key(key_name):
+    try:
+        key = open(key_name, 'r')
+        contents = key.read()
+        return contents
+    except FileNotFoundError as e:
+        logger.error('Could not locate key file', error=e)
+        return
+
+EQ_PUBLIC_KEY = get_key(os.getenv('EQ_PUBLIC_KEY', "/keys/sdc-submission-signing-sr-public-key.pem"))
+EQ_PRIVATE_KEY = get_key(os.getenv('EQ_PRIVATE_KEY', "/keys/sdc-submission-signing-sr-private-key.pem"))
+EQ_PRIVATE_KEY_PASSWORD = os.getenv("EQ_PRIVATE_KEY_PASSWORD", "digitaleq")
+
+PRIVATE_KEY = get_key(os.getenv('PRIVATE_KEY', "/keys/sdc-submission-encryption-sdx-private-key.pem"))
+PRIVATE_KEY_PASSWORD = os.getenv("PRIVATE_KEY_PASSWORD", "digitaleq")